telegram_bot/requirements.txt
python-telegram-bot==22.0
python-dotenv==1.1.0
httpx==0.28.1
anyio==4.9.0
certifi==2025.1.31
httpcore==1.0.7
idna==3.10
h11==0.14.0
sniffio==1.3.1
typing_extensions==4.13.2 